Corrected the calculation of the VDV.
The VDV is the fourth root of the time integral of the acceleration to
the fourth power of the signal. It is not normalized to duration, so we
store the VDV of the raw signal for the first 10 seconds of the
repetition. If the repetition is not 10 seconds long we store NaN.